+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.0.3] - 2026-09-13
 
 ### Fixed
 - Command lines beginning with `quit` that are not exactly `quit\r\n`
@@ -142,6 +142,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added non-root user `USER beanstalk` to run the daemon with unprivileged permissions.
 - Removed deprecated `MAINTAINER` instruction from `Dockerfile`.
 
+[1.0.3]: https://github.com/jonbaldie/beanstalkd/compare/v1.0.2...v1.0.3
 [1.0.2]: https://github.com/jonbaldie/beanstalkd/compare/v1.0.1...v1.0.2
 [1.0.1]: https://github.com/jonbaldie/beanstalkd/compare/v1.0.0...v1.0.1
 [1.0.0]: https://github.com/jonbaldie/beanstalkd/releases/tag/v1.0.0